+++ web/src/resources/product/__tests__/CreateProduct.test.jsx
@@ -1,11 +1,11 @@
-// CreateProductWithRestriction.test.jsx
+// CreateProduct.test.jsx
 
 import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { MemoryRouter, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
-import CreateProductWithRestriction from './CreateProductWithRestriction';
+import CreateProduct from '../views/CreateProduct';
 import WaitOn from '../../../global/components/helpers/WaitOn';
 import * as productService from '../productService'; // We'll mock this
 import { mockUseCreateProduct, mockProduct } from '../__mocks__/ProductService';
@@ -14,9 +14,9 @@ import { initStore } from '../../../config/store';
 
 const { copy } = WaitOn;
 
-// Mock the service hook
+// We mock the service hook
 jest.mock('../productService', () => ({
-  useCreateProductWithRequiredParams: jest.fn(),
+  useCreateProduct: jest.fn(),
 }));
 
 // Prepare a store
@@ -25,41 +25,43 @@ const store = initStore({
   username: 'loggedInUsername@test.com'
 });
 
-// Helper to render CreateProductWithRestriction with provided hook overrides
+// Helper to render CreateProduct with provided hook overrides
 const renderComponent = (overrides = {}) => {
-  productService.useCreateProductWithRequiredParams.mockReturnValue(
+  productService.useCreateProduct.mockReturnValue(
     mockUseCreateProduct(overrides)
   );
 
   return render(
     <Provider store={store}>
-      <MemoryRouter initialEntries={['/products/new-with-restriction']}>
-        <Route path="/products/new-with-restriction">
-          <CreateProductWithRestriction />
+      <MemoryRouter initialEntries={['/products/new']}>
+        <Route path="/products/new">
+          <CreateProduct />
         </Route>
       </MemoryRouter>
     </Provider>
   );
 };
 
+
 // Basic Rendering
-test('renders CreateProductWithRestriction with a "New Super Fancy Product" title', () => {
+test('renders CreateProduct with a "New Product" title', () => {
   renderComponent();
 
-  expect(screen.getByText(/New Super Fancy Product/i)).toBeInTheDocument();
+  expect(screen.getByText(/New Product/i)).toBeInTheDocument();
 });
 
 // Loading State
 test('displays a loading message while creating product is in progress', () => {
+  // Set up a state that indicates loading
   const loadingOverrides = {
-    isLoading: true,
-    isFetching: true,
-    isSuccess: false,
-    data: null
+    isLoading: true
+    , isFetching: true
+    , isSuccess: false
+    , data: null
   };
   renderComponent(loadingOverrides);
 
-  expect(screen.getByText(/Loading\.../i)).toBeInTheDocument();
+  expect(screen.getByText(/Loading\.\.\./i)).toBeInTheDocument();
 });
 
 // Error State
@@ -74,6 +76,7 @@ test('displays server error message when creation query fails', () => {
 
   renderComponent(errorOverrides);
 
+  // WaitOn by default shows a short "An error occurred" + a "Refetch" button
   expect(screen.getByText(/Server error/i)).toBeInTheDocument();
 });
 
@@ -83,11 +86,12 @@ test('displays generic error message if an error is flagged but no error text is
     error: null,
     isLoading: false,
     isFetching: false,
-    isSuccess: false
+    isSuccess: false,
   };
 
   renderComponent(errorOverrides);
 
+  // WaitOn has a default text for unknown errors
   expect(screen.getByText(copy.fetchError.trim())).toBeInTheDocument();
 });
 
@@ -113,7 +117,7 @@ test('displays a message when product data is empty (no default item)', () => {
     isEmpty: true,
     isLoading: false,
     isFetching: false,
-    isSuccess: true
+    isSuccess: true,
   };
 
   renderComponent(emptyOverrides);
@@ -123,7 +127,9 @@ test('displays a message when product data is empty (no default item)', () => {
 
 // Form Interaction
 test('renders the product form with default data', () => {
-  renderComponent({ data: { title: '', description: '', featured: false } });
+  // By default, the mock product is used as initial data in the mock
+  // or you can pass your own data override
+  renderComponent({data: { title: '', description: '', featured: false }});
   const titleInput = screen.getByLabelText(/Title/i);
   const descriptionInput = screen.getByLabelText(/Description/i);
   const featuredCheckbox = screen.getByLabelText(/Featured/i);
@@ -141,8 +147,7 @@ test('calls handleSubmit when form is submitted', () => {
   const handleSubmitMock = jest.fn((e) => {
     e.preventDefault();
     return e.target;
-  });
-
+  })
   renderComponent({ handleSubmit: handleSubmitMock });
 
   const form = screen.getByRole('form');
@@ -151,14 +156,3 @@ test('calls handleSubmit when form is submitted', () => {
   expect(handleSubmitMock).toHaveBeenCalled();
 });
 
-// Server Restriction
-test('ensures requiredParam is sent to the server', () => {
-  const requiredParam = 'super-fancy';
-  renderComponent({
-    requiredParam,
-  });
-
-  expect(productService.useCreateProductWithRequiredParams).toHaveBeenCalledWith(
-    expect.objectContaining({ requiredParam })
-  );
-});
