+++ web/src/resources/product/__tests__/ProductList.test.jsx
@@ -1,32 +1,30 @@
-// MyProducts.test.jsx
+// ProductList.test.jsx
 
 import React from 'react';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import { MemoryRouter, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
-import MyProducts from './MyProducts';
-import { mockUseProductListByLoggedIn, mockUseProductFromMap, mockUseUpdateProductByLoggedInUser } from '../__mocks__/ProductService'; // Mock data
+import ProductList from '../views/ProductList';
+import { mockUseGetProductList, mockUseProductFromMap } from '../__mocks__/ProductService'; // Mock data
 import * as productService from '../productService'; // Mock service
 import { initStore } from '../../../config/store';
 
 jest.mock('../productService', () => ({
-  useProductListByLoggedIn: jest.fn(),
+  useGetProductList: jest.fn(),
   useProductFromMap: jest.fn(),
-  useUpdateProductByLoggedInUser: jest.fn(),
 }));
 
 const store = initStore({ _id: 'loggedInUserId', username: 'loggedInUsername@test.com' });
 
 const renderComponent = (overrides = {}) => {
-  productService.useProductListByLoggedIn.mockReturnValue(mockUseProductListByLoggedIn(overrides));
+  productService.useGetProductList.mockReturnValue(mockUseGetProductList(overrides));
   productService.useProductFromMap.mockImplementation(mockUseProductFromMap);
-  productService.useUpdateProductByLoggedInUser.mockReturnValue(mockUseUpdateProductByLoggedInUser());
   return render(
     <Provider store={store}>
-      <MemoryRouter initialEntries={['/products/mine']}>
-        <Route path="/products/mine">
-          <MyProducts />
+      <MemoryRouter initialEntries={['/products']}>
+        <Route path="/products">
+          <ProductList />
         </Route>
       </MemoryRouter>
     </Provider>
@@ -34,26 +32,23 @@ const renderComponent = (overrides = {}) => {
 };
 
 // Basic Rendering
-test('renders the "My Products" list correctly', () => {
+test('renders the product list correctly', () => {
   renderComponent();
-  // expect(screen.getByText(/My Products/i)).toBeInTheDocument();
-  // My Products is a link in the nav and the page title. Need to test by element type specifically
-  expect(screen.getByRole('heading', { name: /My Products/i })).toBeInTheDocument();
-  expect(screen.getByText(/New Product With Restrictions/i)).toBeInTheDocument();
+  expect(screen.getByText(/Product List/i)).toBeInTheDocument();
+  expect(screen.getByText(/New Product/i)).toBeInTheDocument();
 });
 
 // Loading State
 test('displays the `pagination.per` number of loading skeleton items while fetching data', () => {
   renderComponent({ isLoading: true, isFetching: true, data: null, page: 1, per: 10 });
-  const skeletonListItems = screen.getAllByText(/This is my sample product description/i);
+  const skeletonListItems = screen.getAllByText(/This is a sample product description/i);
   expect(skeletonListItems.length).toBe(10);
 });
 
 // Empty State
 test('displays a message when no products are found', () => {
   renderComponent({ data: [], isEmpty: true });
-  expect(screen.getByText(/No Products Found/i)).toBeInTheDocument();
-  expect(screen.getByText(/You have not created any products yet/i)).toBeInTheDocument();
+  expect(screen.getByText(/No data found/i)).toBeInTheDocument();
 });
 
 // Error State
@@ -82,9 +77,9 @@ test('renders a list of products when data is available', () => {
   expect(screen.getByText(/Product 2/i)).toBeInTheDocument();
 });
 
-// "New Product With Restrictions" Navigation
-test('navigates to the "New Product With Restrictions" page when button is clicked', () => {
+// "New Product" Navigation
+test('navigates to the "New Product" page when button is clicked', () => {
   renderComponent();
-  const newProductLink = screen.getByText(/New Product With Restrictions/i);
-  expect(newProductLink).toHaveAttribute('href', '/products/new-with-restriction');
+  const newProductLink = screen.getByText(/New Product/i);
+  expect(newProductLink).toHaveAttribute('href', '/products/new');
 });
