+++ web/src/resources/product/__tests__/UpdateProduct.test.jsx
@@ -1,14 +1,14 @@
-// CreateProduct.test.jsx
+// UpdateProduct.test.jsx
 
 import React from 'react';
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
+import { render, screen, fireEvent } from '@testing-library/react';
 import { MemoryRouter, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
-import CreateProduct from './CreateProduct';
+import UpdateProduct from '../views/UpdateProduct';
 import WaitOn from '../../../global/components/helpers/WaitOn';
 import * as productService from '../productService'; // We'll mock this
-import { mockUseCreateProduct, mockProduct } from '../__mocks__/ProductService';
+import { mockUseGetUpdatableProduct, mockProduct } from '../__mocks__/ProductService';
 
 import { initStore } from '../../../config/store';
 
@@ -16,44 +16,43 @@ const { copy } = WaitOn;
 
 // We mock the service hook
 jest.mock('../productService', () => ({
-  useCreateProduct: jest.fn(),
+  useGetUpdatableProduct: jest.fn(),
 }));
 
 // Prepare a store
 const store = initStore({
   _id: 'loggedInUserId',
-  username: 'loggedInUsername@test.com'
+  username: 'loggedInUsername@test.com',
 });
 
-// Helper to render CreateProduct with provided hook overrides
+// Helper to render UpdateProduct with provided hook overrides
 const renderComponent = (overrides = {}) => {
-  productService.useCreateProduct.mockReturnValue(
-    mockUseCreateProduct(overrides)
+  productService.useGetUpdatableProduct.mockReturnValue(
+    mockUseGetUpdatableProduct(overrides)
   );
 
   return render(
     <Provider store={store}>
-      <MemoryRouter initialEntries={['/products/new']}>
-        <Route path="/products/new">
-          <CreateProduct />
+      <MemoryRouter initialEntries={[`/products/edit/${mockProduct._id}`]}>
+        <Route path="/products/edit/:id">
+          <UpdateProduct />
         </Route>
       </MemoryRouter>
     </Provider>
   );
 };
 
-
 // Basic Rendering
-test('renders CreateProduct with a "New Product" title', () => {
+test('renders UpdateProduct with the product title in the header', () => {
   renderComponent();
 
-  expect(screen.getByText(/New Product/i)).toBeInTheDocument();
+  expect(screen.getByText(new RegExp('Update Existing Product', 'i'))).toBeInTheDocument();
 });
 
 // Loading State
-test('displays a loading message while creating product is in progress', () => {
-  // Set up a state that indicates loading
-  const loadingOverrides = {
+test('displays a loading message while fetching product data is in progress', () => {
+   // Set up a state that indicates loading
+   const loadingOverrides = {
     isLoading: true
     , isFetching: true
     , isSuccess: false
@@ -61,11 +60,11 @@ test('displays a loading message while creating product is in progress', () => {
   };
   renderComponent(loadingOverrides);
 
-  expect(screen.getByText(/Loading\.\.\./i)).toBeInTheDocument();
+  expect(screen.getByText(/Loading\.+/i)).toBeInTheDocument();
 });
 
 // Error State
-test('displays server error message when creation query fails', () => {
+test('displays server error message when update query fails', () => {
   const errorOverrides = {
     isError: true,
     error: 'Server error',
@@ -76,11 +75,10 @@ test('displays server error message when creation query fails', () => {
 
   renderComponent(errorOverrides);
 
-  // WaitOn by default shows a short "An error occurred" + a "Refetch" button
   expect(screen.getByText(/Server error/i)).toBeInTheDocument();
 });
 
-test('displays generic error message if an error is flagged but no error text is given', () => {
+test('displays generic error message if an error is flagged but no error text is provided', () => {
   const errorOverrides = {
     isError: true,
     error: null,
@@ -88,10 +86,8 @@ test('displays generic error message if an error is flagged but no error text is
     isFetching: false,
     isSuccess: false,
   };
-
   renderComponent(errorOverrides);
 
-  // WaitOn has a default text for unknown errors
   expect(screen.getByText(copy.fetchError.trim())).toBeInTheDocument();
 });
 
@@ -100,7 +96,7 @@ test('allows refetching on error', () => {
   const errorOverrides = {
     isError: true,
     error: 'Fail',
-    refetch: refetchMock
+    refetch: refetchMock,
   };
 
   renderComponent(errorOverrides);
@@ -126,10 +122,9 @@ test('displays a message when product data is empty (no default item)', () => {
 });
 
 // Form Interaction
-test('renders the product form with default data', () => {
-  // By default, the mock product is used as initial data in the mock
-  // or you can pass your own data override
-  renderComponent({data: { title: '', description: '', featured: false }});
+test('renders the product form with existing product data', () => {
+  renderComponent({ data: mockProduct });
+
   const titleInput = screen.getByLabelText(/Title/i);
   const descriptionInput = screen.getByLabelText(/Description/i);
   const featuredCheckbox = screen.getByLabelText(/Featured/i);
@@ -138,16 +133,17 @@ test('renders the product form with default data', () => {
   expect(descriptionInput).toBeInTheDocument();
   expect(featuredCheckbox).toBeInTheDocument();
 
-  expect(titleInput).toHaveValue('');
-  expect(descriptionInput).toHaveValue('');
-  expect(featuredCheckbox).not.toBeChecked();
+  expect(titleInput).toHaveValue(mockProduct.title);
+  expect(descriptionInput).toHaveValue(mockProduct.description);
+  mockProduct.featured ? expect(featuredCheckbox).toBeChecked : expect(featuredCheckbox).not.toBeChecked();
 });
 
 test('calls handleSubmit when form is submitted', () => {
   const handleSubmitMock = jest.fn((e) => {
     e.preventDefault();
     return e.target;
-  })
+  });
+
   renderComponent({ handleSubmit: handleSubmitMock });
 
   const form = screen.getByRole('form');
@@ -155,4 +151,3 @@ test('calls handleSubmit when form is submitted', () => {
 
   expect(handleSubmitMock).toHaveBeenCalled();
 });
-
